Refactor Strategy logic by adopting Domain-Driven Design principles
Moved `Strategy` logic, actions, and model under `Domain\Strategy`. Removed unused models and files related to `Strategy`. Updated references and adjusted documentation and static analysis configuration accordingly.
