+++ docs/tasks.md
@@ -6,6 +6,9 @@ This document provides a comprehensive checklist of improvement tasks for the So
 
 [x] **Implement Domain-Driven Design (DDD) principles**
    - [x] Reorganize code into domain-specific modules
+     - [x] Forecasting
+     - [x] Strategy
+     - [ ] Energy
    - [ ] Create clear boundaries between different domains (e.g., Forecasting, Strategy, Energy Import/Export)
    - [ ] Define value objects for domain concepts
